// models/ui-types.ts
export type Status = "default" | "success" | "error";
export type TextAlignment = "start" | "end" | "center";
export type ToastPosition =
    | "top-left"
    | "top-center"
    | "top-right"
    | "left-center"
    | "right-center"
    | "bottom-left"
    | "bottom-center"
    | "bottom-right";
export type LinkType =
    | "default"
    | "icon"
    | "circleIcon"
    | "underline"
    | "highlighted";
export type LinkSize = "large" | "medium" | "small";
export type LinkColor = "black" | "white";